backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-nmap==0.7.1
httpx==0.27.0
pydantic==2.10.4
python-dotenv==1.0.1
fpdf2==2.7.9
